requirements/base.txt
# Core
Django==6.0.4
psycopg2-binary==2.9.11
djangorestframework==3.16.0

# Multi-tenancy
django-tenants==3.10.1

# Auth
django-allauth==65.7.0
django-axes==7.0.1
django-guardian==2.4.0

# Async
celery[redis]==5.5.1
redis==5.2.1
django-celery-beat==2.9.0

# Observability
sentry-sdk[django]==2.25.1
django-structlog==9.1.0

# Security & Encryption
django-ratelimit==4.1.0
django-encrypted-model-fields==0.6.5

# Billing
stripe==12.1.0

# Utils
django-environ==0.12.0
gunicorn==23.0.0
